:building_construction: Changed the architecture to be deployed on vercel
