+++ backend/src/api/index.ts
@@ -1,0 +1,16 @@
+import express from 'express';
+
+import MessageResponse from '../interfaces/MessageResponse';
+import emojis from './emojis';
+
+const router = express.Router();
+
+router.get<{}, MessageResponse>('/', (req, res) => {
+  res.json({
+    message: 'API - 👋🌎🌍🌏',
+  });
+});
+
+router.use('/emojis', emojis);
+
+export default router;
