+++ backend/src/interfaces/ErrorResponse.ts
@@ -1,0 +1,5 @@
+import MessageResponse from './MessageResponse';
+
+export default interface ErrorResponse extends MessageResponse {
+  stack?: string;
+}
