+++ backend/src/interfaces/MessageResponse.ts
@@ -1,0 +1,3 @@
+export default interface MessageResponse {
+  message: string;
+}
